requirements.txt
kurigram
openai
Pillow
python-magic
rlottie-python
TgCrypto
google-generativeai
ffmpeg
PyYAML
beautifulsoup4
aiohttp